Impact of the occult on state legitimacy and democratization aid in Africa
The impact of occult belief on legitimacy of the state and on aid for democratization in Africa: Among politicians and development experts in Africa alike there is a growing awareness of the never decreasing importance of the belief in magic and witchcraft on political decision making since pre-colonial times. Demonstration of the control of occult forces as a means of enhancing legitimacy of traditional or charismatic rule had been considered for over a century to be the prerogative of traditional chiefs and their marabouts; now it proved to be effective for the modern political elite and the state as well. An increasing number of African states officially recognized the existence of magic and witchcraft, and they adapted the imposed colonial law accordingly. In addition magic-religious belief systems, as represented by the vodun or African independent churches (e.g. the Kimbanguists), boasting to control witchcraft, are promoted by African political leaders to enhance legitimacy both of the political class and of state governance. At the same time, development experts tried to take into account the socio cultural dimension of development; they called for an "endogenization" of development aid. This call was justified, because endogenization should be considered as a pre-requisite of sustainable aid; however, under certain conditions it may be ambiguous and dangerous as well. Concerning the consideration of occult belief it may lead to the promotion of illegitimate rule and violation of basic human rights.