Simbabwe : Bestandsaufnahme der Förderpolitik und -praxis der evangelischen Hilfswerke EZE / BfdW
The development policy concepts of the ‘Brot für die Welt’ (BfdW) and ‘Evangelische Zentralstelle für Entwicklungshilfe’ (EZE) exemplary, both in terms of their content and in terms of their formulation, in comparison with other non-governmental organizations (NGOs) worldwide. A conceptual weak point, however, is their rather ecclesiastical postulate, based on general church policies, i.e. that the pastoral mandate on the one side and the charitable or development policy mandate on the other could easily be reconciled without compromise. In the practice of development cooperation, this postulate leads to the dogmatisation of the choice of the local partner overseas, irrespectively of their qualification. Yet, exactly this form of labour-division is demanded by most of the donor- and partner institutions in pursuit of their shared principle of equal partnership and their maxim of helping the people to help themselves.